Fix the breaking circle specs
*Something* in the *local* (i.e. two of my dev machines and DS's
machine) test stacks is rescuing the NoMethodError that would otherwise
be raised by `#authorise!(code)` when the parsed JSON response does not
have the expected keys. The same thing does not happen on CircleCI. It
does, however, happen if you interrupt the local test with a
`binding.pry` and call the statement by hand.

Returning valid demo data fixes this locally and **should** fix it on
Circle.