Moj sso oauth2 (#4)
* First cut integration with MoJ SSO

This has been run and checked against a local version of the MoJ
SSO. I am anticipating additional work will be required when it comes time
to deploy it against the real server. It may also be the case that the
workflow needs refining

Mount new routes and get basic workflow

Setting up mocks.

Ignore .env files

* Update dev instruction for moj-sso

* Stop circle from breaking

* Add a pointer to the ENV docs

* Fix the breaking circle specs

*Something* in the *local* (i.e. two of my dev machines and DS's
machine) test stacks is rescuing the NoMethodError that would otherwise
be raised by `#authorise!(code)` when the parsed JSON response does not
have the expected keys. The same thing does not happen on CircleCI. It
does, however, happen if you interrupt the local test with a
`binding.pry` and call the statement by hand.

Returning valid demo data fixes this locally and **should** fix it on
Circle.