Enable logging (#9)
This outputs it in logstash json format to STDOUT to make it easier to
integrate with kibana. I also replaced the generic login link with a
message directing the user to log in via an existing case.  This was
done to prevent users getting locked into a login/logout loop (it logs
them straight out if they have not requested a specific set of
documents).