Persist OAuth authentication to S3 bucket
OAuth can tell us if a user has a login, but we still need to persist
this to the app.  This solves the problem by using another S3 bucket to
handle the persistence.  The idea is that the bucket would be configured
to have a (suitably short–probably 1 day) lifecycle on new objects. This
would require users to sign in via MoJ-SSO every 24 hours.