Reduce brittleness of specs
I chose to use VCR when I was uncertain about the complexity and
specifics of using the AWS SDK to deal with S3.  In the intervening
months, my understanding–and the understanding of the team–has improved.
The VCR spec dependency, and the chore of VCR cassette maintenance for
new features in particular, now feels like more of a burden than a
benefit, so I am refactoring to remove it.  This is first example.