Fix mutation problems and remove healthcheck
The mutation tests threw up some changes that would not have worked had
they been implmented.  David caught these and I changed them.

I am removing healthcheck at this point as it needs to be renamed to
status, *and* it I am doing this in a two-stage merge given the
complexity of changing the mocking/stubbing approach.  I will leave
this branch open after the spec changes are in place and close it once
status is merged.