Healthcheck (#17)
* Keep well-behaved robots off of the site

It should never be indexed.

* WIP: read test

* Reduce brittleness of specs

I chose to use VCR when I was uncertain about the complexity and
specifics of using the AWS SDK to deal with S3.  In the intervening
months, my understanding–and the understanding of the team–has improved.
The VCR spec dependency, and the chore of VCR cassette maintenance for
new features in particular, now feels like more of a burden than a
benefit, so I am refactoring to remove it.  This is first example.

* Fix mutation problems and remove healthcheck

The mutation tests threw up some changes that would not have worked had
they been implmented.  David caught these and I changed them.

I am removing healthcheck at this point as it needs to be renamed to
status, *and* it I am doing this in a two-stage merge given the
complexity of changing the mocking/stubbing approach.  I will leave
this branch open after the spec changes are in place and close it once
status is merged.